deploy pipelines at startup
