+++ src/hayhooks/cli/run/__init__.py
@@ -1,9 +1,12 @@
 import click
 import uvicorn
+import os
 
 
 @click.command()
-@click.option('-h', '--host', default="localhost")
-@click.option('-p', '--port', default=1416)
-def run(host, port):
+@click.option('--host', default="localhost")
+@click.option('--port', default=1416)
+@click.option('--pipelines-dir', default="pipelines.d")
+def run(host, port, pipelines_dir):
+    os.environ["PIPELINES_DIR"] = pipelines_dir
     uvicorn.run("hayhooks.server:app", host=host, port=port)
